+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2025-02-13
+## [0.2.0] - 2025-04-11
+
+### Added
+
+- Ensure compatibility with older versions that do not support optional chaining.
+- Allow user context to be optionally passed to the `useGetFlag` hook, enabling it to override the context provided to the `VWOProvider` hook.
+
+## [0.1.0] - 2025-02-13
 
 ### Added
 
@@ -13,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
   #### Basic Usage
 
-  ```javascript
+  ```typescript
   import React from 'react';
   import { VWOProvider } from 'vwo-fme-react-sdk';
 
@@ -34,7 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   };
 
   const App = () => (
-    <VWOProvider config={vwoConfig} context={userContext}>
+    <VWOProvider config={vwoConfig} userContext={userContext}>
       <YourComponent />
     </VWOProvider>
   );